+++ InstabugSample/.prettierrc.js
@@ -1,8 +1,0 @@
-module.exports = {
-  arrowParens : 'avoid',
-  bracketSameLine : true,
-  printWidth : 100,
-  singleQuote : true,
-  trailingComma : 'all',
-};
-